add loop detection
