+++ LICENSE
@@ -1,3 +1,11 @@
+Copyright (c) 2020-present Bricks Cloud Technologies, Inc.
+
+Portions of this software are licensed as follows:
+
+* All content that resides under the "ee" directory of this repository, if that directory exists, is licensed under the license defined in "ee" directory.
+* All third party components incorporated into the Bricks Software are licensed under the original license provided by the owner of the applicable component.
+* Content outside of the above mentioned directories or restrictions above is available under the "Apache" license as defined below.
+
                                  Apache License
                            Version 2.0, January 2004
                         http://www.apache.org/licenses/
