+++ README.md
@@ -9,7 +9,7 @@ Bricks is an open source tool for converting Figma designs into high-quality fro
 <p align="center">
    <a href='http://makeapullrequest.com'><img alt='PRs Welcome' src='https://img.shields.io/badge/PRs-welcome-43AF11.svg?style=shields'/></a>
    <a href="https://github.com/bricks-cloud/bricks/stargazers"><img src="https://img.shields.io/github/stars/bricks-cloud/bricks?color=e4b442" alt="Github Stars"></a>
-   <a href="https://join.slack.com/t/brickscommunity/shared_invite/zt-1pb2hy3h2-9rDYWMZdHKxHblzUG0CpTQ"><img src="https://img.shields.io/badge/slack-bricks-blue?logo=slack&labelColor=2EB67D" alt="Join Bricks on Slack"></a>
+   <a href="https://discord.gg/NM6aeBeqCD"><img src="https://img.shields.io/badge/discord-bricks-blue?logo=discord&labelColor=2EB67D" alt="Join Bricks on Discord"></a>
    <a href="https://github.com/bricks-cloud/bricks/blob/main/LICENSE"><img src="https://img.shields.io/badge/license-Apache-red" alt="License"></a>
    <a href="https://github.com/bricks-cloud/bricks/commits/main"><img alt="GitHub commit activity" src="https://img.shields.io/github/commit-activity/m/bricks-cloud/bricks?color=8b55e3"/></a>
 </p>
